Merge Mike's connection refcount work.
The attached patch adds a connection refcount and forces the cleanup of
pending tasks or tasks that are not going to get completed by are not
running in tgtd. The refcout is needed for the latter. When those
commands have completed we then release the connection.

This patch only handles the memory issues. Should we actually not close
the connection endpoints and should we send the result of the commands
that were executing in tgtd when they finally complete? For some like
linux-iscsi or open-iscsi, when the connection is dropped we assume the
command is not executed, but tgtd might actually complete it. Is either
wrong?